Holding a process of memories in a 3-D Bloch of oil
Hi, I'm michaeljmakerMan
I had a couple of items that where new on traks? I could practically build new any subjects & abouts & or! Abouts...whom can record me is a resiviour of oil, that's the amounts of what it'd be to hold my memories,... I'd think it'd haft to be oils Fuji film= which is ***** fish & ****....call me I could build or we could build a platform so much better than others 3-D for certain key times or arranging different times to knowing people voice's recorded knows when someone else is involved or gets involved